agent-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
google-genai>=0.5.0
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pinecone>=5.0.0
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
duckdb>=0.10.0
watchdog>=3.0.0
# Observability
structlog>=24.0.0
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client>=0.19.0
bleach>=6.1.0
